+++ spec/policy/keycloak_scope_check/keycloak_scope_check_spec.lua
@@ -1,10 +1,10 @@
-local ScopeCheckPolicy = require('apicast.policy.scope_check')
+local KeycloakScopeCheckPolicy = require('apicast.policy.keycloak_scope_check')
 
-describe('Scope check policy', function()
+describe('Keycloak Scope check policy', function()
   local ngx_exit_spy
   local ngx_say_spy
 
-  setup(function()
+  before_each(function()
     ngx_exit_spy = spy.on(ngx, 'exit')
     ngx_say_spy = spy.on(ngx, 'say')
   end)
@@ -13,7 +13,7 @@ describe('Scope check policy', function()
     describe('whitelist', function()
       describe('check succeeds', function()
         it('realm role', function()
-          local scope_check_policy = ScopeCheckPolicy.new({
+          local scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
                 realm_roles = { { name = "aaa" } },
@@ -34,11 +34,12 @@ describe('Scope check policy', function()
             }
           }
 
-          assert(scope_check_policy:access(context))
+          scope_check_policy:access(context)
+          assert.spy(ngx_say_spy).was_not_called()
         end)
 
         it('client role', function()
-          local scope_check_policy = ScopeCheckPolicy.new({
+          local scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
                 client_roles = { { name = "aaa", client = "ccc" } },
@@ -61,11 +62,12 @@ describe('Scope check policy', function()
             }
           }
 
-          assert(scope_check_policy:access(context))
+          scope_check_policy:access(context)
+          assert.spy(ngx_say_spy).was_not_called()
         end)
 
         it('multi roles in policy', function()
-          local scope_check_policy = ScopeCheckPolicy.new({
+          local scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
                 realm_roles = { { name = "ddd" } },
@@ -92,59 +94,61 @@ describe('Scope check policy', function()
             }
           }
 
-          assert(scope_check_policy:access(context))
+          scope_check_policy:access(context)
+          assert.spy(ngx_say_spy).was_not_called()
         end)
 
         it('wildcard', function()
-          local scope_check_policy = ScopeCheckPolicy.new({
+          local scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
-                client_roles = { { name = "aaa", client = "ccc" } },
-                resource = "/bbb/{wildcard}"
+                client_roles = { { name = "aaa", client = "client" } },
+                resource = "/{wildcard}/client"
               }
             }
           })
 
           ngx.var = {
-            uri = '/bbb/ddd'
+            uri = '/group-10/client/resources'
           }
 
           local context = {
             jwt = {
               resource_access = {
-                ccc = {
-                  roles = { "aaa", "eee" }
+                client = {
+                  roles = { "aaa", "other_role" }
                 }
               }
             }
           }
 
-          assert(scope_check_policy:access(context))
+          scope_check_policy:access(context)
+          assert.spy(ngx_say_spy).was_not_called()
         end)
 
         it('multi scopes', function()
-          local scope_check_policy = ScopeCheckPolicy.new({
+          local scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
-                client_roles = { { name = "aaa", client = "ccc" } },
-                resource = "/bbb/{wildcard}"
+                client_roles = { { name = "role_of_known_client", client = "known_client" } },
+                resource = "/not-accessed/"
               },
               {
-                realm_roles = { { name = "rrr" } },
-                client_roles = { { name = "fff", client = "hhh" } },
-                resource = "/eee{wildcard}"
+                realm_roles = { { name = "unknown_role" } },
+                client_roles = { { name = "role_of_known_client", client = "unknown_client" } },
+                resource = "/account-a"
               },
               {
-                realm_roles = { { name = "ddd" } },
-                resource = "/eeeyyy{wildcard}"
+                realm_roles = { { name = "known_role" } },
+                resource = "/account-a"
               },
               {
-                realm_roles = { { name = "rrr" } },
-                resource = "/{wildcard}ggg"
+                realm_roles = { { name = "unknown_role" } },
+                resource = "/{wildcard}/account-b"
               },
               {
-                client_roles = { { name = "fff", client = "hhh" } },
-                resource = "/z{wildcard}ggg"
+                client_roles = { { name = "role_of_known_client", client = "known_client" } },
+                resource = "/group-{wildcard}/account-b"
               }
             }
           })
@@ -152,11 +156,11 @@ describe('Scope check policy', function()
           local context = {
             jwt = {
               realm_access = {
-                roles = { "ddd" }
+                roles = { "known_role" }
               },
               resource_access = {
-                hhh = {
-                  roles = { "fff" }
+                known_client = {
+                  roles = { "role_of_known_client" }
                 }
               }
             },
@@ -167,16 +171,18 @@ describe('Scope check policy', function()
           }
 
           ngx.var = {
-            uri = '/eeeyyyy'
+            uri = '/account-a'
           }
 
-          assert(scope_check_policy:access(context))
+          scope_check_policy:access(context)
+          assert.spy(ngx_say_spy).was_not_called()
 
           ngx.var = {
-            uri = '/zzzggg'
+            uri = '/group-a/account-b'
           }
 
-          assert(scope_check_policy:access(context))
+          scope_check_policy:access(context)
+          assert.spy(ngx_say_spy).was_not_called()
         end)
       end)
 
@@ -185,19 +191,19 @@ describe('Scope check policy', function()
         local context
 
         before_each(function()
-          scope_check_policy = ScopeCheckPolicy.new({
+          scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
-                client_roles = { { name = "aaa", client = "ccc" } },
-                resource = "/bbb/{wildcard}"
+                client_roles = { { name = "role_of_known_client", client = "known_client" } },
+                resource = "/match"
               },
               {
-                realm_roles = { { name = "ddd" }, { name = "uuu" } },
-                resource = "/eee{wildcard}"
+                client_roles = { { name = "role_of_known_client", client = "unknown_client" } },
+                resource = "/no-role-resource"
               },
               {
-                client_roles = { { name = "fff", client = "hhh" } },
-                resource = "/{wildcard}ggg"
+                realm_roles = { { name = "known_role" }, { name = "unknown_role" } },
+                resource = "/not-enough-roles-resource"
               },
             }
           })
@@ -205,11 +211,11 @@ describe('Scope check policy', function()
           context = {
             jwt = {
               realm_access = {
-                roles = { "ddd" }
+                roles = { "known_role" }
               },
               resource_access = {
-                hhh = {
-                  roles = { "fff" }
+                known_client = {
+                  roles = { "role_of_known_client" }
                 }
               }
             },
@@ -222,7 +228,7 @@ describe('Scope check policy', function()
 
         it('not match the uri', function()
           ngx.var = {
-            uri = '/zzz/gggaaa'
+            uri = '/not-match'
           }
 
           scope_check_policy:access(context)
@@ -233,7 +239,7 @@ describe('Scope check policy', function()
 
         it('no role', function()
           ngx.var = {
-            uri = '/bbb/xxx'
+            uri = '/no-role-resource'
           }
 
           scope_check_policy:access(context)
@@ -244,7 +250,7 @@ describe('Scope check policy', function()
 
         it('not enough roles', function()
           ngx.var = {
-            uri = '/eeeaaa'
+            uri = '/not-enough-roles-resource'
           }
 
           scope_check_policy:access(context)
@@ -258,7 +264,7 @@ describe('Scope check policy', function()
     describe('blacklist', function()
       describe('check fails', function()
         it('realm role', function()
-          local scope_check_policy = ScopeCheckPolicy.new({
+          local scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
                 realm_roles = { { name = "aaa" } },
@@ -291,7 +297,7 @@ describe('Scope check policy', function()
         end)
 
         it('client role', function()
-          local scope_check_policy = ScopeCheckPolicy.new({
+          local scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
                 client_roles = { { name = "aaa", client = "ccc" } },
@@ -326,7 +332,7 @@ describe('Scope check policy', function()
         end)
 
         it('multi roles in policy', function()
-          local scope_check_policy = ScopeCheckPolicy.new({
+          local scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
                 realm_roles = { { name = "ddd" } },
@@ -365,25 +371,25 @@ describe('Scope check policy', function()
         end)
 
         it('wildcard', function()
-          local scope_check_policy = ScopeCheckPolicy.new({
+          local scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
-                client_roles = { { name = "aaa", client = "ccc" } },
-                resource = "/bbb/{wildcard}"
+                client_roles = { { name = "aaa", client = "client" } },
+                resource = "/{wildcard}/client"
               }
             },
             type = "blacklist"
           })
 
           ngx.var = {
-            uri = '/bbb/ddd'
+            uri = '/group-10/client/resources'
           }
 
           local context = {
             jwt = {
               resource_access = {
-                ccc = {
-                  roles = { "aaa", "eee" }
+                client = {
+                  roles = { "aaa", "other_role" }
                 }
               }
             },
@@ -400,28 +406,28 @@ describe('Scope check policy', function()
         end)
 
         it('multi scopes', function()
-          local scope_check_policy = ScopeCheckPolicy.new({
+          local scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
-                client_roles = { { name = "aaa", client = "ccc" } },
-                resource = "/bbb/{wildcard}"
+                client_roles = { { name = "role_of_known_client", client = "known_client" } },
+                resource = "/not-accessed/"
               },
               {
-                realm_roles = { { name = "rrr" } },
-                client_roles = { { name = "fff", client = "hhh" } },
-                resource = "/eee{wildcard}"
+                realm_roles = { { name = "unknown_role" } },
+                client_roles = { { name = "role_of_known_client", client = "unknown_client" } },
+                resource = "/account-a"
               },
               {
-                realm_roles = { { name = "ddd" } },
-                resource = "/eeeyyy{wildcard}"
+                realm_roles = { { name = "known_role" } },
+                resource = "/account-a"
               },
               {
-                realm_roles = { { name = "rrr" } },
-                resource = "/{wildcard}ggg"
+                realm_roles = { { name = "unknown_role" } },
+                resource = "/group-{wildcard}/account-b"
               },
               {
-                client_roles = { { name = "fff", client = "hhh" } },
-                resource = "/z{wildcard}ggg"
+                client_roles = { { name = "role_of_known_client", client = "known_client" } },
+                resource = "/group-{wildcard}/account-b"
               }
             },
             type = "blacklist"
@@ -430,11 +436,11 @@ describe('Scope check policy', function()
           local context = {
             jwt = {
               realm_access = {
-                roles = { "ddd" }
+                roles = { "known_role" }
               },
               resource_access = {
-                hhh = {
-                  roles = { "fff" }
+                known_client = {
+                  roles = { "role_of_known_client" }
                 }
               }
             },
@@ -445,7 +451,7 @@ describe('Scope check policy', function()
           }
 
           ngx.var = {
-            uri = '/eeeyyyy'
+            uri = '/account-a'
           }
 
           scope_check_policy:access(context)
@@ -454,7 +460,7 @@ describe('Scope check policy', function()
           assert.spy(ngx_say_spy).was_called_with("auth failed")
 
           ngx.var = {
-            uri = '/zzzggg'
+            uri = '/group-a/account-b'
           }
 
           scope_check_policy:access(context)
@@ -469,19 +475,19 @@ describe('Scope check policy', function()
         local context
 
         before_each(function()
-          scope_check_policy = ScopeCheckPolicy.new({
+          scope_check_policy = KeycloakScopeCheckPolicy.new({
             scopes = {
               {
-                client_roles = { { name = "aaa", client = "ccc" } },
-                resource = "/bbb/{wildcard}"
+                client_roles = { { name = "role_of_known_client", client = "known_client" } },
+                resource = "/match"
               },
               {
-                realm_roles = { { name = "ddd" }, { name = "uuu" } },
-                resource = "/eee{wildcard}"
+                client_roles = { { name = "role_of_known_client", client = "unknown_client" } },
+                resource = "/no-role-resource"
               },
               {
-                client_roles = { { name = "fff", client = "hhh" } },
-                resource = "/{wildcard}ggg"
+                realm_roles = { { name = "known_role" }, { name = "unknown_role" } },
+                resource = "/not-enough-roles-resource"
               },
             },
             type = "blacklist"
@@ -490,11 +496,11 @@ describe('Scope check policy', function()
           context = {
             jwt = {
               realm_access = {
-                roles = { "ddd" }
+                roles = { "known_role" }
               },
               resource_access = {
-                hhh = {
-                  roles = { "fff" }
+                known_client = {
+                  roles = { "role_of_known_client" }
                 }
               }
             },
@@ -507,26 +513,29 @@ describe('Scope check policy', function()
 
         it('not match the uri', function()
           ngx.var = {
-            uri = '/zzz/gggaaa'
+            uri = '/not-match'
           }
 
-          assert(scope_check_policy:access(context))
+          scope_check_policy:access(context)
+          assert.spy(ngx_say_spy).was_not_called()
         end)
 
         it('no role', function()
           ngx.var = {
-            uri = '/bbb/xxx'
+            uri = '/no-role-resource'
           }
 
-          assert(scope_check_policy:access(context))
+          scope_check_policy:access(context)
+          assert.spy(ngx_say_spy).was_not_called()
         end)
 
         it('not enough roles', function()
           ngx.var = {
-            uri = '/eeeaaa'
+            uri = '/not-enough-roles-resource'
           }
 
-          assert(scope_check_policy:access(context))
+          scope_check_policy:access(context)
+          assert.spy(ngx_say_spy).was_not_called()
         end)
       end)
     end)